http-redirect: Minimal awk script to redirect http to https
A simple script written in awk to parse HTTP/1.1 headers and then
generate a reply to use the https:// URL.

If and only if the request method  is GET or HEAD will
the path URL will be appended.

Recgonises multi-line headers, repeated headers, and absolute URI
(where the host is in the URI and the Host: header is ignored).

Deisgned to run by systemd socket activation.

Tested: test cases in ./tests show location is parsed
	Tested on a linux box with gawk
	Tested via nc -l -p 8080 exec awf -f redirect.awk & curl localhost:8080
	Unit tests scripts
	Second pass at writing systemd activation (untested, closer to console)

Change-Id: Ic361247dedfd7bcff620d793dd7082c247c9e78e
Signed-off-by: Milton Miller <miltonm@us.ibm.com>