http-redirect: Minimal awk script to redirect http to https
A simple script written in awk to parse HTTP/1.1 headers and then
generate a reply to use the https:// URL.

If and only if the request method  is GET or HEAD will
the path URL will be appended.

Recgonises multi-line headers, repeated headers, and absolute URI
(where the host is in the URI and the Host: header is ignored).
Does not (yet) decode %xx url encoding.

Includes systemd socket and service files.

Tested: test cases in ./tests show location is parsed
	Tested on a linux box with gawk
	Tested via nc -l -p 8080 -c awk -f redirect.awk &
		curl -v http://localhost:8080/some/path

	Tested on withspoon with awk script in /home/root/redirect.awk
         * ran Unit tests script
       	 * systemd units installed in /run/systemd/system/
         * curl -w %{redirect_url}'\n' http://hostname.example/

Change-Id: Ic361247dedfd7bcff620d793dd7082c247c9e78e
Signed-off-by: Milton Miller <miltonm@us.ibm.com>