http-redirect: Simplify installing service files
The output file name for the service files was unintentionally
different than the input file name. By installing the service
files in a loop, it gets rid of duplicate code and allows the
use of a variable for the input and output field, fixing the typo.

If a variable was not used, output could be set to @PLAINNAME@
to use the same name as input.

Change-Id: Ic0cef9a21227bbd712a0214c6e78814ba27860ab
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>