usb-ctrl: factor out function sysfs path and simplify quoting
The function name in sysfs was assembled several times when configuring
a mass storage device.  Change usb_set_interface_type to not expect the
path to have a trailing slash.  Consolidate the function name to one
variable and use it.

Quote complete strings instead of individual variables reducing the
number of quotes needed.  Quotes are not required when a safe string is
assigned within the script.

Change-Id: I6c070957b1b9b36693af5717566d9bb13e93269d
Signed-off-by: Milton Miller <miltonm@us.ibm.com>