usb-ctrl: Add support to set BMC's mac address
By default it uses the random MAC address when usb ecm is enabled.
Inspired from usb-network.sh in meta-quanta, it is possible to specify
the MAC address.

Add an optional argument to set the MAC address when ecm is enabled, so
that the BMC usb0 has known MAC address.

Tested: Verify the usb0 MAC address could be specified by below command:

            usb-ctrl ecm usbeth0 on <mac-addr>

Signed-off-by: Lei YU <yulei.sh@bytedance.com>
Change-Id: Ib726b070a08a06dcff9ba0110925af1aed4c77cc