usb-ctrl: Add support to set BMC and Host's mac address
By default it uses the random MAC address when usb ecm is enabled.
Inspired from usb-network.sh in meta-quanta, it is possible to specify
the MAC address.

Add optional arguments to set the MAC address when ecm is enabled, so
that the BMC and Host side's MAC addresses are known.

Tested: Verify the MAC addresses could be specified by below command:

            usb-ctrl ecm usbeth0 on <bmc-addr host-addr>

Signed-off-by: Lei YU <yulei.sh@bytedance.com>
Signed-off-by: Milton Miller <miltonm@us.ibm.com>
Change-Id: I4a944d62d3bbb7e57d1beab90e435663d3ff1e1f