meson: set fake language
The organization-wide CI currently assumes that every meson-based
project is either C or C++ and attempts to run sanitizer options
which are only applicable to those languages.  This causes the CI
to fail on this repository.  This repository only uses meson.build
as a way to direct scripts for install and requires no compile or
language, but set a fake language of 'c' to bypass the CI issues.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Ifcaefbe75f955df8bc54e5c3ac1aee27ee88f6e5