first-boot-set-hostname: Restart mDNS services when the hostname changes
mDNS services do not immediately broadcast the updated name when
the hostname is changed. As a result, the system remains discoverable
only under the previous hostname. To address this issue, the relevant
mDNS services are restarted when a new hostname is applied, ensuring
that the new hostname is advertised correctly.

Change-Id: I39f2392609e2ef883546df13089b6ea980199b5e
Signed-off-by: Jae Hyun Yoo <jae.yoo@oss.qualcomm.com>